Admin MS Updated project
 Updated with Loan entity ,controller,Repository,service class

+++ Admin_User_creation/Admin_User_Creation/src/main/java/com/swiz/bcs/repository/AdminRepository.java
@@ -3,5 +3,5 @@ package com.swiz.bcs.repository;
 import com.swiz.bcs.entity.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<UserAccount,Long> {
+public interface AdminRepository extends JpaRepository<UserAccount,Integer> {
 }
